controlador atualizado

+++ trunk/Emprestimus/src/controlador/EmprestimusIF.java
@@ -5,7 +5,7 @@ public interface EmprestimusIF {
 	
 	public void criarUsuario( String login, String nome, String endereco );
 	
-	public int abrirSessao( String login );
+	public Long abrirSessao( String login );
 	
 	public String getAtributoUsuario( String login, String atributo );
